requirements.txt
flask>=3.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
newspaper3k>=0.2.8
gnews>=0.3.7
googlenewsdecoder>=0.1.7
googlesearch-python>=1.2.0
lxml>=4.9.0
lxml_html_clean>=0.1.0
gunicorn>=22.0.0
